+++ src/main/java/com/jandi/plan_backend/commu/dto/CommunityItemDTO.java
@@ -1,7 +1,7 @@
 package com.jandi.plan_backend.commu.dto;
 
 import com.jandi.plan_backend.commu.entity.Community;
-import com.jandi.plan_backend.storage.service.ImageService;
+import com.jandi.plan_backend.image.service.ImageService;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
