+++ src/main/java/com/jandi/plan_backend/commu/dto/ParentCommentDTO.java
@@ -1,7 +1,7 @@
 package com.jandi.plan_backend.commu.dto;
 
 import com.jandi.plan_backend.commu.entity.Comment;
-import com.jandi.plan_backend.storage.service.ImageService;
+import com.jandi.plan_backend.image.service.ImageService;
 import com.jandi.plan_backend.user.entity.User;
 import lombok.Getter;
