+++ src/main/java/com/jandi/plan_backend/user/controller/ProfileImageController.java
@@ -1,6 +1,6 @@
 package com.jandi.plan_backend.user.controller;
 
-import com.jandi.plan_backend.storage.dto.ImageResponseDto;
+import com.jandi.plan_backend.image.dto.ImageResponseDto;
 import com.jandi.plan_backend.security.CustomUserDetails;
 import com.jandi.plan_backend.user.service.ProfileImageService;
 import lombok.extern.slf4j.Slf4j;
